standardising boosts to Double

+++ elastic4s-core/src/main/scala/com/sksamuel/elastic4s/searches/queries/ExistsQueryDefinition.scala
@@ -1,11 +1,11 @@
 package com.sksamuel.elastic4s.searches.queries
 
 case class ExistsQueryDefinition(field: String,
-                                 boost: Option[Float] = None,
+                                 boost: Option[Double] = None,
                                  queryName: Option[String] = None) extends QueryDefinition {
 
-  def boost(boost: Float): ExistsQueryDefinition = copy(boost = Option(boost))
-  def withBoost(boost: Float): ExistsQueryDefinition = copy(boost = Option(boost))
+  def boost(boost: Double): ExistsQueryDefinition = copy(boost = Option(boost))
+  def withBoost(boost: Double): ExistsQueryDefinition = copy(boost = Option(boost))
 
   def queryName(queryName: String): ExistsQueryDefinition = copy(queryName = Option(queryName))
   def withQueryName(queryName: String): ExistsQueryDefinition = copy(queryName = Option(queryName))
