+++ elastic4s-tcp/src/main/scala/com/sksamuel/elastic4s/searches/queries/BuildableTermsQueryImplicits.scala
@@ -22,8 +22,8 @@ trait BuildableTermsQueryImplicits {
       QueryBuilders.termsQuery(q.field, q.values.toSeq: _*)
   }
 
-  implicit object FloatBuildableTermsQuery extends BuildableTermsQuery[Float] {
-    override def build(q: TermsQueryDefinition[Float]): TermsQueryBuilder =
+  implicit object FloatBuildableTermsQuery extends BuildableTermsQuery[Double] {
+    override def build(q: TermsQueryDefinition[Double]): TermsQueryBuilder =
       QueryBuilders.termsQuery(q.field, q.values.toSeq: _*)
   }
